pyproject.toml
[project]
name = "htls"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "idna>=3.10",
    "pyhumps>=3.8.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
"htls" = ["libs/*"]
